Moved common components to shared folder
