+++ src/app/dataset-view/additional-components/data-component/data.component.spec.ts
@@ -12,20 +12,20 @@ import { RouterTestingModule } from "@angular/router/testing";
 import { Location } from "@angular/common";
 import { FormsModule } from "@angular/forms";
 import { MatDividerModule } from "@angular/material/divider";
-import { LoadMoreComponent } from "./load-more/load-more.component";
+import { LoadMoreComponent } from "../../../query/shared/load-more/load-more.component";
 import { DynamicTableModule } from "../../../components/dynamic-table/dynamic-table.module";
 import { EditorModule } from "src/app/shared/editor/editor.module";
 import { MatProgressBarModule } from "@angular/material/progress-bar";
 import { OverviewUpdate } from "../../dataset.subscriptions.interface";
 import _ from "lodash";
-import { RequestTimerComponent } from "./request-timer/request-timer.component";
+import { RequestTimerComponent } from "../../../query/shared/request-timer/request-timer.component";
 import { SqlEditorComponent } from "src/app/shared/editor/components/sql-editor/sql-editor.component";
 import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
 import { HttpClientModule } from "@angular/common/http";
 import { Apollo } from "apollo-angular";
 import { ToastrModule } from "ngx-toastr";
-import { SavedQueriesSectionComponent } from "./saved-queries-section/saved-queries-section.component";
-import { QueryAndResultSectionsComponent } from "./query-and-result-sections/query-and-result-sections.component";
+import { SavedQueriesSectionComponent } from "../../../query/shared/saved-queries-section/saved-queries-section.component";
+import { QueryAndResultSectionsComponent } from "../../../query/shared/query-and-result-sections/query-and-result-sections.component";
 
 describe("DataComponent", () => {
     let component: DataComponent;
