+++ src/app/query/global-query/global-query.component.spec.ts
@@ -1,16 +1,16 @@
-import { SavedQueriesSectionComponent } from "./../../dataset-view/additional-components/data-component/saved-queries-section/saved-queries-section.component";
+import { SavedQueriesSectionComponent } from "../shared/saved-queries-section/saved-queries-section.component";
 import { DynamicTableModule } from "./../../components/dynamic-table/dynamic-table.module";
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { GlobalQueryComponent } from "./global-query.component";
 import { EditorModule } from "src/app/shared/editor/editor.module";
 import { CdkAccordionModule } from "@angular/cdk/accordion";
-import { QueryAndResultSectionsComponent } from "src/app/dataset-view/additional-components/data-component/query-and-result-sections/query-and-result-sections.component";
+import { QueryAndResultSectionsComponent } from "src/app/query/shared/query-and-result-sections/query-and-result-sections.component";
 import { Apollo } from "apollo-angular";
 import { HttpClientTestingModule } from "@angular/common/http/testing";
 import { MatMenuModule } from "@angular/material/menu";
 import { MatProgressBarModule } from "@angular/material/progress-bar";
 import { ToastrModule } from "ngx-toastr";
-import { RequestTimerComponent } from "src/app/dataset-view/additional-components/data-component/request-timer/request-timer.component";
+import { RequestTimerComponent } from "src/app/query/shared/request-timer/request-timer.component";
 import { ActivatedRoute } from "@angular/router";
 import ProjectLinks from "src/app/project-links";
 import { MatIconModule } from "@angular/material/icon";
