+++ src/app/query/shared/load-more/load-more.component.scss
