+++ src/app/query/shared/query-and-result-sections/query-and-result-sections.component.html
