+++ src/app/query/shared/saved-queries-section/saved-queries-section.component.scss
